Improved code and started to implement a PMM
 - Started implementing a PMM
 - Now GNUmakefile actually uses a cross compiler
 - Improved code
 - Updated README.md

+++ kernel/arch/x86_64/idt.c
@@ -1,4 +1,5 @@
 
+#include <stdint.h>
 #include <arch/x86_64/include/idt.h>
 
 
@@ -19,7 +20,7 @@ static void idtSetEntry(uint8_t i, void* isrAddr, uint8_t ist, uint8_t flags) {
 	idt[i].ist = ist;
 	idt[i].flags = flags;
 	idt[i].isr1 = ((uint64_t)isrAddr >> 16) & 0xFFFFFFFFFFFF;
-	idt[i].reserved = (uint32_t)0;
+	idt[i].reserved = 0;
 }
